\name{plot.hsmm.data}
\alias{plot.hsmm.data}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Plot function for hsmm data }
\description{
  Produces a plot of the observed sequences, and displays a coloured bar signifying the hidden states (if available)
}
\usage{
\method{plot}{hsmm.data}(x, ...)
}

\arguments{
  \item{x}{ A \code{hsmm.data} object}
  \item{\dots}{ Arguments passed to \code{plot.ts} }
}
\author{Jared O'Connell}