\name{plot.hsmm}
\alias{plot.hsmm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Plot function for hsmms }
\description{
  Displays the densities for the sojourn distributions of each state.
}
\usage{
\method{plot}{hsmm}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ A \code{hsmm} object }
  \item{\dots}{ Arguments passed to \code{plot}}
}
\author{Jared O'Connell jaredoconnell@gmail.com}