\name{print.hmm}
\alias{print.hmm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print method for hmm objects }
\description{
  Prints the slots of a hmm object
}
\usage{
\method{print}{hmm}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ An object of type \code{hmm}}
  \item{\dots}{further arguments passed to or from other methods.}
}
\author{Jared O'Connell jaredoconnell@gmail.com}
