\name{print.hmmspec}
\alias{print.hmmspec}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print function for hmmspec }
\description{
  Prints the parameters contained in the object
}
\usage{
\method{print}{hmmspec}(x, ...)
}

\arguments{
  \item{x}{ An object of type \code{hmmspec}}
  \item{\dots}{further arguments passed to or from other methods.}

}
\author{Jared O'Connell jaredoconnell@gmail.com}
