\name{print.hsmmspec}
\alias{print.hsmmspec}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print function for hsmmspec }
\description{
  Prints the parameters contained in the object
}
\usage{
\method{print}{hsmmspec}(x, ...)
}

\arguments{
  \item{x}{ An object of type \code{hsmmspec}}
  \item{\dots}{further arguments passed to or from other methods.}

}
\author{Jared O'Connell jaredoconnell@gmail.com}
