\name{reproai}
\docType{data}
\alias{reproai}
\title{Artificial insemination times for seven cows}
\description{
  This is an auxilliary data set to the \code{cows} data set containing times of artificial insemination for respective cows.  
  Only the day of insemination was recorded so time of day is always midday.
}
\usage{reproai}
\format{\code{reproai} is a dataframe with 12 rows and \code{id} being the cow's id and \code{days.from.calving} recording the number of days from calving when insemination occurred.
}
\source{Danish Cattle Research Centre}
\references{
Peters, A. and Ball, P. (1995), "Reproduction in Cattle," 2nd ed.
}
\keyword{datasets}
