\name{reprocows}
\docType{data}
\alias{reprocows}
\title{Reproductive data from seven dairy cows}
\description{
  This data set contains hourly observations on progesterone and an activity index at hourly intervals since calving on seven dairy cows.  
}
\usage{reprocows}
\format{\code{reprocows} is a data frame containing 13040 rows. 
\code{id} is the cow ID, \code{progesterone} is a measurement of the hormone in ng/L taken from a milk sample, activity is a relative measure of activity calculated from a pedometer.

There are a large number of missing values as progesterone is measured only at milking time (and at a farm manager's discretion).  Missing values in activity occur due to hardware problems can occur with pedometers.
}
\source{Danish Cattle Research Centre}
\references{
Peters, A. and Ball, P. (1995), "Reproduction in Cattle," 2nd ed.
}
\keyword{datasets}
