\name{reproppa}
\docType{data}
\alias{reproppa}
\title{Observed lengths of post-partum anoestrus for 73 dairy cows}
\description{
  This data set contains the observed length of post-partum anoestrus (in days) for 73 dairy cattle.
}
\usage{reproppa}
\format{\code{reproppa} a vector containing 73 integers.
}
\source{Danish Cattle Research Centre}
\references{
Peters, A. and Ball, P. (1995), "Reproduction in Cattle," 2nd ed.
}
\keyword{datasets}
