\name{simulate}
\alias{simulate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Generic simualation funciton in the mhsmm package }
\description{
  Simulates data from a specified mhsmm model
}


\usage{
simulate(object,...)
}

\arguments{
  \item{object}{ An object to simulate from}
  \item{...}{ Arguments to be passed to methods. }
}


\author{ Jared O'Connell jaredoconnell@gmail.com }


\seealso{ 
  \code{\link{simulate.hsmmspec}},
  \code{\link{simulate.hmmspec}}
}