\name{summary.hmm}
\alias{summary.hmm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Summary method for hmm objects }
\description{
  Prints the estimated parameters of a \code{hmm} object
}
\usage{
\method{summary}{hmm}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ A \code{hmm} object }
  \item{\dots}{further arguments passed to or from other methods.}

}

\value{
  An object of class 'summary.hmm'
}

\author{Jared O'Connell jaredoconnell@gmail.com}
