\name{summary.hsmm}
\alias{summary.hsmm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Summary function for hsmm }
\description{
  Returns a summary object for a hsmm object
}
\usage{
\method{summary}{hsmm}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ An object of type \code{hsmm} }
  \item{\dots}{further arguments passed to or from other methods.}

}
\author{Jared O'Connell jaredoconnell@gmail.com}